+++ OpenRA.Mods.Common/Traits/Sound/DeathSounds.cs
@@ -42,7 +42,7 @@ namespace OpenRA.Mods.Common.Traits
 
 			if (info.DeathTypes.Contains(e.Warhead.DeathType) || (!info.DeathTypes.Any() &&
 				!self.Info.Traits.WithInterface<DeathSoundsInfo>().Any(dsi => dsi.DeathTypes.Contains(e.Warhead.DeathType))))
-				self.PlayVoiceLocal(self, info.DeathSound, self.Owner.Country.Race, info.VolumeMultiplier);
+				self.PlayVoiceLocal(info.DeathSound, info.VolumeMultiplier);
 		}
 	}
 }
